requirements.txt
# AI Avatar Platform - Python Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Deep Learning & AI
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1

# Computer Vision
opencv-python==4.8.1.78
mediapipe==0.10.8
face-recognition==1.3.0
dlib==19.24.2
Pillow==10.1.0

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
scipy==1.11.4
numpy==1.24.3

# Voice Synthesis
elevenlabs==0.2.27
pydub==0.25.1

# Video Processing
moviepy==1.0.3
imageio==2.33.0
imageio-ffmpeg==0.4.9

# HuggingFace & Model Management
huggingface-hub==0.19.4
transformers==4.35.2
accelerate==0.25.0

# ML Utilities
scikit-learn==1.3.2
scikit-image==0.22.0

# API & Web
requests==2.31.0
aiohttp==3.9.1
aiofiles==23.2.1
httpx==0.25.2

# Database & Caching
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.0

# Job Queue
celery==5.3.4
rq==1.15.1

# Monitoring & Logging
sentry-sdk==1.38.0
loguru==0.7.2
prometheus-client==0.19.0

# Utilities
tqdm==4.66.1
colorama==0.4.6
click==8.1.7
pyyaml==6.0.1
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Rate Limiting
slowapi==0.1.9

# CORS
fastapi-cors==0.0.6

# File Handling
python-magic==0.4.27

# Progress Tracking
websockets==12.0

# Environment Detection
psutil==5.9.6
GPUtil==1.4.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.5.2
